phan quyen + change pass

+++ src/components/User/UserItem.js
@@ -12,24 +12,51 @@ class UserItem extends Component {
             this.props.deleteUser(id);
         }
     }
+
+    renderTr = (user, userData) => {
+        if (userData.role == 'boss' && (user.role == "user" || user.role == "admin" || user.id == userData.id)) {
+            return (
+                <td>
+                    <Link to title="Sửa" className="edit"><i className="fa fa-pencil icon" /></Link>
+                    <Link to='/' title="Xóa" className="delete" onClick={(e, id) => this.deleteClick(e, user.id)}><i className="fa fa-trash icon" /></Link>
+                </td>
+            )
+        } else {
+            if (userData.role == 'admin' && (user.role == "user" || user.id == userData.id)) {
+                return (
+                    <td>
+                        <Link to title="Sửa" className="edit"><i className="fa fa-pencil icon" /></Link>
+                        <Link to='/' title="Xóa" className="delete" onClick={(e, id) => this.deleteClick(e, user.id)}><i className="fa fa-trash icon" /></Link>
+                    </td>
+                )
+            }
+            else {
+                return (
+                    <td>
+                    </td>
+                )
+            }
+        }
+
+
+    }
     render() {
         let { user } = this.props;
+        var userData = localStorage.getItem("userData");
+        userData = JSON.parse(userData);
         return (
             <tr>
                 <td>{this.props.stt}</td>
                 <td>{user.name}</td>
                 <td>{user.email}</td>
                 <td>{user.role}</td>
-                <td>
-                    <Link to title="Sửa" className="edit"><i className="fa fa-pencil icon" /></Link>
-                    <Link to='/' title="Xóa" className="delete" onClick={(e, id) => this.deleteClick(e, user.id)}><i className="fa fa-trash icon" /></Link>
-                </td>
+                {this.renderTr(user, userData)}
             </tr>
         );
     }
 }
 
-const mapDispatchToProps = (dispatch, ownProps) => {
+const mapDispatchToProps = (dispatch) => {
     return {
         deleteUser: (id) => {
             dispatch(actDeleteUserRequest(id))
